move to sparsesets on component_manager

+++ src/ecs/entity.rs
@@ -1,2 +1,2 @@
 pub type Entity = usize;
-pub type Signature = u8;
+pub type Signature = u16;
